refactor: setup gh actions and added thinking trace (#1)
