+++ .github/actions/setup/action.yml
@@ -1,0 +1,10 @@
+name: Setup UV Action
+description: Set up UV for use in workflows
+
+runs:
+  using: composite
+  steps:
+    - name: Install uv
+      uses: astral-sh/setup-uv@v3
+      with:
+        version: "0.8.0"
